fix: Service table backend setup
