+++ src/routes/dashboard/admin/services/+page.server.ts
@@ -1,6 +1,6 @@
 import { Service } from "$lib/classes/Service";
 
 export async function load() {
-	/* Loads student records from the DB when page is created. */
+	/* Loads service records from the DB when page is created. */
 	return Service.selectServices();
 }
